# openapi.yaml
openapi: 3.0.2
info:
  title: Recruitment CRM API (RESTful)
  version: 2.0.0
  description: >
    REST API системы управления наймом.
    Полностью соответствует модели данных (сущности + связи).
    Все GET-запросы поддерживают регистронезависимый поиск по подстроке.
servers:
  - url: http://localhost:8000/api
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TestTaskCreate:
      type: object
      description: "Тестовое задание, прикрепляется к вакансии"
      properties:
        title: { type: string, maxLength: 100, example: "Анализ требований к API" }
        test_task_url: { type: string, format: uri, description: "Ссылка на условия задания" }
        vacancy_id: { type: string, format: uuid, description: "Связь TEST_FOR: TestTask → Vacancy" }
      required: [title, test_task_url, vacancy_id]
    TestTaskResponse:
      type: object
      description: "Тестовое задание, прикрепляется к вакансии"
      properties:
        id: { type: string, format: uuid, description: "task_id" }
        title: { type: string, maxLength: 100, example: "Анализ требований к API" }
        test_task_url: { type: string, format: uri, description: "Ссылка на условия задания" }
        vacancy_id: { type: string, format: uuid, description: "Связь TEST_FOR: TestTask → Vacancy" }
      required: [id, title, test_task_url, vacancy_id]
...
paths:
...
  /v2/test-tasks:
    get:
      summary: Список тестовых заданий
      tags: [TestTasks]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: title
          in: query
          schema: { type: string }
          description: Поиск по подстроке
        - name: vacancy_id
          in: query
          schema: { type: string, format: uuid }
        - name: vacancy_title
          in: query
          schema: { type: string }
          description: Поиск по названию вакансии (связанная сущность)
        - name: test_task_url_contains
          in: query
          schema: { type: string }
          description: Поиск по ссылке (подстрока)
        - name: has_assigned_candidates
          in: query
          schema: { type: boolean }
          description: Есть ли кандидаты с этим тестовым
        - name: limit
          in: query
          schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
        - name: offset
          in: query
          schema: { type: integer, default: 0, minimum: 0 }
        - name: sort_by
          in: query
          schema: { type: string, enum: [title, vacancy_id, created_at], default: title }
        - name: sort_order
          in: query
          schema: { type: string, enum: [asc, desc], default: asc }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { type: integer }
                  items: { type: array, items: { $ref: '#/components/schemas/TestTaskResponse' } }

    post:
      summary: Создать тестовое задание для вакансии
      tags: [TestTasks]
      description: Роль HR. Тестовое задание должно быть связано с вакансией (vacancy_id), для которой оно предназначено (связь TEST_FOR). В одной вакансии может быть несколько тестовых заданий
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TestTaskCreate' }
      responses:
        '201':
             description: Создано тестовое
             content:
                application/json: { schema: { $ref: '#/components/schemas/TestTaskResponse' } }
        '404': { description: Вакансия не найдена }
  /v2/test-tasks/{id}:
    get:
      summary: Получить тестовое задание по ID
      description: Роль - любая
      tags: [TestTasks]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
            description: OK
            content: { application/json: { schema: { $ref: '#/components/schemas/TestTaskResponse' } } }
        '404': { description: Не найдено }
    patch:
      summary: Обновить тестовое задание
      description: Роль HR, ADMIN может обновлять тестовые задания.
      tags: [TestTasks]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string, maxLength: 100 }
                test_task_url: { type: string, format: uri }
      responses:
        '200':
            description: OK
            content: { application/json: { schema: { $ref: '#/components/schemas/TestTaskResponse' } } }
        '404': { description: Не найдено }
    delete:
      summary: Удалить тестовое задание
      description: Роль HR, ADMIN может удалять тестовые задания. Удаление тестового задания должно автоматически снимать его со всех кандидатов, которым оно было назначено (связь COMPLETES)
      tags: [TestTasks]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Удалено }
        '404': { description: Не найдено }

  # ==================== Candidates ====================
...